refactor: split dce try pruning
